fix: change presets
